% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_check_behavior_summary.R
\name{mb_check_behavior_summary}
\alias{mb_check_behavior_summary}
\title{Check consistency of behaviour summary}
\usage{
mb_check_behavior_summary(behavior_summary, behaviors, t = 6)
}
\arguments{
\item{behavior_summary}{Data.frame of summaries (rows = mice).}

\item{behaviors}{Character vector of behaviour base names.}

\item{t}{Test length in minutes.}
}
\value{
Invisibly returns the data.frame; issues warnings on problems.
}
\description{
\itemize{
\item Entry1 should not be after the test duration.
\item If Entry1 is NA, Time and all Min columns should be 0.
}
}
