% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_check_columns.R
\name{mb_check_columns}
\alias{mb_check_columns}
\title{Check that a CowLog file has required columns}
\usage{
mb_check_columns(file, filename = NA_character_)
}
\arguments{
\item{file}{A data.frame.}

\item{filename}{Optional filename for messages.}
}
\value{
The same data.frame (invisibly), or an error if columns are missing.
}
\description{
Check that a CowLog file has required columns
}
