% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_fill_na_summary.R
\name{mb_fill_na_summary}
\alias{mb_fill_na_summary}
\title{Fill NA values in summary according to forced swim rules}
\usage{
mb_fill_na_summary(behavior_summary)
}
\arguments{
\item{behavior_summary}{Data.frame of behaviour summaries.}
}
\value{
Data.frame with NA values replaced.
}
\description{
\itemize{
\item all *Min columns: NA -> 0
\item all *Time columns: NA -> 0
\item all *Entry1 columns: NA -> -9
}
}
