% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_mouse_id_from_path.R
\name{mb_mouse_id_from_path}
\alias{mb_mouse_id_from_path}
\title{Derive MouseID from file path}
\usage{
mb_mouse_id_from_path(path)
}
\arguments{
\item{path}{File path.}
}
\value{
Character MouseID.
}
\description{
Example: "Mouse1_ForcedSwim.csv" -> "Mouse1"
}
