% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_summarise_file.R
\name{mb_summarise_file}
\alias{mb_summarise_file}
\title{Summarise all behaviours for a single file}
\usage{
mb_summarise_file(file, behaviors, t = 6, mouse_id = NULL)
}
\arguments{
\item{file}{Cleaned CowLog data.frame for one mouse.}

\item{behaviors}{Character vector of behaviour base names.}

\item{t}{Test length in minutes.}

\item{mouse_id}{Optional MouseID; if NULL, you can set it later.}
}
\value{
One-row data.frame with columns:
MouseID, \if{html}{\out{<b>}}Time, \if{html}{\out{<b>}}Entry1, \if{html}{\out{<b>}}Bouts, \if{html}{\out{<b>}}Min1..MinT.
}
\description{
Summarise all behaviours for a single file
}
